moving old ocropus functionality to new file for rewrite
